requirements.txt
pyserial==3.5
Flask==2.3.3
Flask-CORS==4.0.0